Clarify dataset loader comment
